[executors] fix singleplayer
