+++ libraries/executors/executors-mc12w18a-mc1.7.5/gradle.properties
@@ -1,0 +1,6 @@
+min_mc_version = 12w18a
+max_mc_version = 1.7.5
+minecraft_dependency = >=1.3-alpha.12.18.a <=1.7.5
+
+minecraft_version = 1.7.5
+nests_build = 5
