+++ libraries/executors/executors-mc12w18a-mc1.7.5/src/main/java/net/ornithemc/osl/executors/impl/mixin/common/MinecraftServerMixin.java
@@ -1,11 +1,14 @@
-package net.ornithemc.osl.executors.impl.mixin.server;
+package net.ornithemc.osl.executors.impl.mixin.common;
 
 import java.util.ArrayDeque;
 import java.util.Queue;
 
+import org.objectweb.asm.Opcodes;
+
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.At.Shift;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
@@ -52,7 +55,10 @@ public class MinecraftServerMixin implements MainThreadExecutor {
 	@Inject(
 		method = "tick",
 		at  = @At(
-			value = "HEAD"
+			value = "FIELD",
+			target = "Lnet/minecraft/server/MinecraftServer;ticks:I",
+			opcode = Opcodes.PUTFIELD,
+			shift = Shift.AFTER
 		)
 	)
 	private void osl$executors$runPendingTasks(CallbackInfo ci) {
@@ -75,15 +81,5 @@ public class MinecraftServerMixin implements MainThreadExecutor {
 		} catch (Throwable t) {
 			Executors.LOGGER.fatal("Error running task", t);
 		}
-	}
-
-	@Inject(
-		method = "shutdown",
-		at = @At(
-			value = "TAIL"
-		)
-	)
-	private void osl$executors$shutdownBackgroundExecutor(CallbackInfo ci) {
-		Executors.shutdownBackgroundExecutor();
 	}
 }
