+++ libraries/executors/executors-mc12w18a-mc1.7.5/src/main/resources/fabric.mod.json
@@ -1,0 +1,26 @@
+{
+  "schemaVersion": 1,
+  "id": "osl-executors",
+  "version": "0.1.0-alpha.1+mc12w18a-mc1.7.5",
+  "environment": "*",
+  "mixins": [
+    "osl.executors.mixins.json"
+  ],
+  "accessWidener": "osl.executors.classtweaker",
+  "depends": {
+    "fabricloader": "\u003e\u003d0.17.3",
+    "minecraft": "\u003e\u003d1.3-alpha.12.18.a \u003c\u003d1.7.5",
+    "osl-core": "\u003e\u003d0.7.0"
+  },
+  "name": "OSL Executors",
+  "description": "Interface for main thread and background threads executors.",
+  "authors": [
+    "OrnitheMC"
+  ],
+  "contact": {
+    "homepage": "https://ornithemc.net/",
+    "issues": "https://github.com/OrnitheMC/ornithe-standard-libraries/issues",
+    "sources": "https://github.com/OrnitheMC/ornithe-standard-libraries"
+  },
+  "license": "Apache-2.0"
+}
