+++ libraries/executors/executors-mc14w02a-mc14w06b/src/main/resources/osl.executors.mixins.json
@@ -4,6 +4,7 @@
 	"package": "net.ornithemc.osl.executors.impl.mixin",
 	"compatibilityLevel": "JAVA_8",
 	"mixins": [
+		"common.MinecraftServerMixin"
 	],
 	"client": [
 		"client.MinecraftMixin"
