+++ libraries/executors/executors-mc14w07a-mc14w20b/src/main/resources/osl.executors.mixins.json
@@ -4,6 +4,7 @@
 	"package": "net.ornithemc.osl.executors.impl.mixin",
 	"compatibilityLevel": "JAVA_8",
 	"mixins": [
+		"common.MinecraftServerMixin"
 	],
 	"client": [
 		"client.MinecraftMixin"
