+++ libraries/executors/executors-mc14w21a-mc1.13.2/src/main/java/net/ornithemc/osl/executors/impl/mixin/common/MinecraftServerMixin.java
@@ -1,0 +1,17 @@
+package net.ornithemc.osl.executors.impl.mixin.common;
+
+import java.util.concurrent.Executor;
+
+import org.spongepowered.asm.mixin.Mixin;
+
+import net.minecraft.server.MinecraftServer;
+import net.minecraft.util.BlockableEventLoop;
+
+@Mixin(MinecraftServer.class)
+public class MinecraftServerMixin implements Executor {
+
+	@Override
+	public void execute(Runnable command) {
+		((BlockableEventLoop) this).executeTask(command);
+	}
+}
