+++ libraries/executors/executors-mc14w21a-mc1.13.2/src/main/java/net/ornithemc/osl/executors/impl/mixin/server/MinecraftServerMixin.java
@@ -1,24 +1,16 @@
 package net.ornithemc.osl.executors.impl.mixin.server;
 
-import java.util.concurrent.Executor;
-
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
 import net.minecraft.server.MinecraftServer;
-import net.minecraft.util.BlockableEventLoop;
 
 import net.ornithemc.osl.executors.impl.Executors;
 
 @Mixin(MinecraftServer.class)
-public class MinecraftServerMixin implements Executor {
-
-	@Override
-	public void execute(Runnable command) {
-		((BlockableEventLoop) this).executeTask(command);
-	}
+public class MinecraftServerMixin {
 
 	@Inject(
 		method = "shutdown",
