--- libraries/executors/executors-mcb1.9-pre3-mc1.7.5/gradle.properties
@@ -1,6 +1,0 @@
-min_mc_version = b1.9-pre3-1350
-max_mc_version = 1.7.5
-minecraft_dependency = >=1.0.0-beta.9.0.3 <=1.7.5
-
-minecraft_version = 1.7.5
-nests_build = 5
