+++ libraries/executors/executors-mcb1.9-pre3-mc12w17a/build.gradle
@@ -1,0 +1,1 @@
+setUpModule(project)
