+++ libraries/executors/executors-mcb1.9-pre3-mc12w17a/gradle.properties
@@ -1,0 +1,7 @@
+min_mc_version = b1.9-pre3-1350
+max_mc_version = 12w17a
+minecraft_dependency = >=1.0.0-beta.9.0.3 <=1.3-alpha.12.17.a
+
+minecraft_version = 12w17a
+client_nests_build = 10
+server_nests_build = 5
