+++ libraries/executors/executors-mcb1.9-pre3-mc12w17a/src/main/resources/osl.executors.classtweaker
@@ -1,0 +1,4 @@
+classTweaker	v1	named
+
+transitive-inject-interface	net/minecraft/client/Minecraft	net/ornithemc/osl/executors/api/MainThreadExecutor
+transitive-inject-interface	net/minecraft/server/MinecraftServer	net/ornithemc/osl/executors/api/MainThreadExecutor
