+++ settings.gradle
@@ -41,7 +41,8 @@ include ':libraries:entrypoints:entrypoints-mc13w16a-mc1.14.4'
 
 include ':libraries:executors'
 include ':libraries:executors:executors-mcin-20100128-2304-mcb1.9-pre2'
-include ':libraries:executors:executors-mcb1.9-pre3-mc1.7.5'
+include ':libraries:executors:executors-mcb1.9-pre3-mc12w17a'
+include ':libraries:executors:executors-mc12w18a-mc1.7.5'
 include ':libraries:executors:executors-mc1.7.6-pre1-mc1.7.10'
 include ':libraries:executors:executors-mc14w02a-mc14w06b'
 include ':libraries:executors:executors-mc14w07a-mc14w20b'
